+++ include/envoy/router/router.h
@@ -151,7 +151,8 @@ public:
   static const uint32_t RETRY_ON_GRPC_DEADLINE_EXCEEDED  = 0x40;
   static const uint32_t RETRY_ON_GRPC_RESOURCE_EXHAUSTED = 0x80;
   static const uint32_t RETRY_ON_GRPC_UNAVAILABLE        = 0x100;
-  static const uint32_t RETRY_ON_RETRIABLE_STATUS_CODES  = 0x200;
+  static const uint32_t RETRY_ON_GRPC_INTERNAL           = 0x200;
+  static const uint32_t RETRY_ON_RETRIABLE_STATUS_CODES  = 0x400;
   // clang-format on
 
   virtual ~RetryPolicy() {}
@@ -172,13 +173,15 @@ public:
   virtual uint32_t retryOn() const PURE;
 
   /**
-   * Initializes the RetryHostPredicates to be used with this retry attempt.
+   * Initializes a new set of RetryHostPredicates to be used when retrying with this retry policy.
    * @return list of RetryHostPredicates to use
    */
   virtual std::vector<Upstream::RetryHostPredicateSharedPtr> retryHostPredicates() const PURE;
 
   /**
-   * @return the RetryPriority to use when determining priority load for retries.
+   * Initializes a RetryPriority to be used when retrying with this retry policy.
+   * @return the RetryPriority to use when determining priority load for retries, or nullptr
+   * if none should be used.
    */
   virtual Upstream::RetryPrioritySharedPtr retryPriority() const PURE;