+++ include/envoy/stream_info/stream_info.h
@@ -9,7 +9,9 @@
 #include "envoy/common/time.h"
 #include "envoy/http/protocol.h"
 #include "envoy/stream_info/filter_state.h"
-#include "envoy/upstream/upstream.h"
+#include "envoy/upstream/host_description.h"
+
+#include "common/protobuf/protobuf.h"
 
 #include "absl/types/optional.h"
 
@@ -306,12 +308,13 @@ public:
 
   /**
    * Object on which filters can share data on a per-request basis.
-   * Only one filter can produce a named data object, but it may be
-   * consumed by many other objects.
-   * @return the per-request state associated with this request.
+   * For singleton data objects, only one filter can produce a named data object.
+   * List data objects can be updated by multiple filters (append only). Both object
+   * types can be consumed by multiple filters.
+   * @return the filter state associated with this request.
    */
-  virtual FilterState& perRequestState() PURE;
-  virtual const FilterState& perRequestState() const PURE;
+  virtual FilterState& filterState() PURE;
+  virtual const FilterState& filterState() const PURE;
 
   /**
    * @param SNI value requested