+++ source/extensions/filters/network/sni_cluster/sni_cluster.cc
@@ -17,7 +17,7 @@ Network::FilterStatus SniClusterFilter::onNewConnection() {
 
   if (!sni.empty()) {
     // Set the tcp_proxy cluster to the same value as SNI
-    read_callbacks_->connection().perConnectionState().setData(
+    read_callbacks_->connection().streamInfo().filterState().setData(
         TcpProxy::PerConnectionCluster::Key, std::make_unique<TcpProxy::PerConnectionCluster>(sni));
   }