+++ test/common/router/router_test.cc
@@ -1735,7 +1735,7 @@ TEST_F(RouterTest, RetryUpstreamGrpcCancelled) {
       }));
   expectResponseTimerCreate();
 
-  Http::TestHeaderMapImpl headers{{"x-envoy-grpc-retry-on", "cancelled"},
+  Http::TestHeaderMapImpl headers{{"x-envoy-retry-grpc-on", "cancelled"},
                                   {"x-envoy-internal", "true"},
                                   {"content-type", "application/grpc"},
                                   {"grpc-timeout", "20S"}};