+++ test/server/options_impl_test.cc
@@ -57,17 +57,6 @@ TEST_F(OptionsImplTest, InvalidCommandLine) {
                           "Couldn't find match for argument");
 }
 
-TEST_F(OptionsImplTest, v1Allowed) {
-  std::unique_ptr<OptionsImpl> options = createOptionsImpl(
-      "envoy --mode validate --concurrency 2 -c hello --admin-address-path path --restart-epoch 1 "
-      "--local-address-ip-version v6 -l info --service-cluster cluster --service-node node "
-      "--service-zone zone --file-flush-interval-msec 9000 --drain-time-s 60 --log-format [%v] "
-      "--parent-shutdown-time-s 90 --log-path /foo/bar --allow-deprecated-v1-api "
-      "--disable-hot-restart");
-  EXPECT_EQ(Server::Mode::Validate, options->mode());
-  EXPECT_FALSE(options->v2ConfigOnly());
-}
-
 TEST_F(OptionsImplTest, v1Disallowed) {
   std::unique_ptr<OptionsImpl> options = createOptionsImpl(
       "envoy --mode validate --concurrency 2 -c hello --admin-address-path path --restart-epoch 1 "
@@ -111,7 +100,6 @@ TEST_F(OptionsImplTest, All) {
 
 TEST_F(OptionsImplTest, SetAll) {
   std::unique_ptr<OptionsImpl> options = createOptionsImpl("envoy -c hello");
-  bool v2_config_only = options->v2ConfigOnly();
   bool hot_restart_disabled = options->hotRestartDisabled();
   Stats::StatsOptionsImpl stats_options;
   stats_options.max_obj_name_length_ = 54321;
@@ -121,7 +109,6 @@ TEST_F(OptionsImplTest, SetAll) {
   options->setConcurrency(42);
   options->setConfigPath("foo");
   options->setConfigYaml("bogus:");
-  options->setV2ConfigOnly(!options->v2ConfigOnly());
   options->setAdminAddressPath("path");
   options->setLocalAddressIpVersion(Network::Address::IpVersion::v6);
   options->setDrainTime(std::chrono::seconds(42));
@@ -143,7 +130,6 @@ TEST_F(OptionsImplTest, SetAll) {
   EXPECT_EQ(42U, options->concurrency());
   EXPECT_EQ("foo", options->configPath());
   EXPECT_EQ("bogus:", options->configYaml());
-  EXPECT_EQ(!v2_config_only, options->v2ConfigOnly());
   EXPECT_EQ("path", options->adminAddressPath());
   EXPECT_EQ(Network::Address::IpVersion::v6, options->localAddressIpVersion());
   EXPECT_EQ(std::chrono::seconds(42), options->drainTime());
@@ -229,6 +215,47 @@ TEST_F(OptionsImplTest, IncompleteComponentLogLevel) {
   std::unique_ptr<OptionsImpl> options = createOptionsImpl("envoy --mode init_only");
   EXPECT_THROW_WITH_REGEX(parseComponentLogLevels(options, "upstream"), MalformedArgvException,
                           "component log level not correctly specified 'upstream'");
+}
+
+// Test that the test constructor comes up with the same default values as the main constructor.
+TEST_F(OptionsImplTest, SaneTestConstructor) {
+  std::unique_ptr<OptionsImpl> regular_options_impl(createOptionsImpl("envoy"));
+  OptionsImpl test_options_impl("service_cluster", "service_node", "service_zone",
+                                spdlog::level::level_enum::info);
+
+  // Specified by constructor
+  EXPECT_EQ("service_cluster", test_options_impl.serviceClusterName());
+  EXPECT_EQ("service_node", test_options_impl.serviceNodeName());
+  EXPECT_EQ("service_zone", test_options_impl.serviceZone());
+  EXPECT_EQ(spdlog::level::level_enum::info, test_options_impl.logLevel());
+
+  // Special (simplified) for tests
+  EXPECT_EQ(1u, test_options_impl.concurrency());
+
+  EXPECT_EQ(regular_options_impl->baseId(), test_options_impl.baseId());
+  EXPECT_EQ(regular_options_impl->configPath(), test_options_impl.configPath());
+  EXPECT_EQ(regular_options_impl->configYaml(), test_options_impl.configYaml());
+  EXPECT_EQ(regular_options_impl->v2ConfigOnly(), test_options_impl.v2ConfigOnly());
+  EXPECT_EQ(regular_options_impl->adminAddressPath(), test_options_impl.adminAddressPath());
+  EXPECT_EQ(regular_options_impl->localAddressIpVersion(),
+            test_options_impl.localAddressIpVersion());
+  EXPECT_EQ(regular_options_impl->drainTime(), test_options_impl.drainTime());
+  EXPECT_EQ(spdlog::level::level_enum::info, test_options_impl.logLevel());
+  EXPECT_EQ(regular_options_impl->componentLogLevels(), test_options_impl.componentLogLevels());
+  EXPECT_EQ(regular_options_impl->logPath(), test_options_impl.logPath());
+  EXPECT_EQ(regular_options_impl->parentShutdownTime(), test_options_impl.parentShutdownTime());
+  EXPECT_EQ(regular_options_impl->restartEpoch(), test_options_impl.restartEpoch());
+  EXPECT_EQ(regular_options_impl->mode(), test_options_impl.mode());
+  EXPECT_EQ(regular_options_impl->fileFlushIntervalMsec(),
+            test_options_impl.fileFlushIntervalMsec());
+  EXPECT_EQ(regular_options_impl->maxStats(), test_options_impl.maxStats());
+  EXPECT_EQ(regular_options_impl->statsOptions().maxNameLength(),
+            test_options_impl.statsOptions().maxNameLength());
+  EXPECT_EQ(regular_options_impl->statsOptions().maxObjNameLength(),
+            test_options_impl.statsOptions().maxObjNameLength());
+  EXPECT_EQ(regular_options_impl->statsOptions().maxStatSuffixLength(),
+            test_options_impl.statsOptions().maxStatSuffixLength());
+  EXPECT_EQ(regular_options_impl->hotRestartDisabled(), test_options_impl.hotRestartDisabled());
 }
 
 } // namespace Envoy